SB26-016 Activate external printer/parser
This commit contains a test for external printer/parser and adds the
necessary attributes to make it work on arrays.

See why3/plugins/ada_terms/README for new easily supported features.

* share/spark/theories/ada__model.mlw
Added syntax and name attributes in array theory to make array access work
in task.

* testsuite/gnatprove/tests/S327-015__itp_desired_tasks/*
Add test for external printer/parser in replay mode. The replay will test
that transformation with arguments as Ada parameters indeed works.

* testsuite/gnatprove/tests/QA11-003__itp_server_test/test.out
This test prints the task so we update it.

Change-Id: Id92628c03c3828625755ab9a06a201255292cca7